promptflow/model/requirements.txt
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-resource
azure-mgmt-subscription
azure-mgmt-compute
azure-mgmt-storage
azure-mgmt-authorization
azure-mgmt-web
azure-mgmt-loganalytics
azure-mgmt-containerinstance
azure-mgmt-containerservice
azure-mgmt-monitor
azure-mgmt-keyvault
azure-mgmt-cognitiveservices
azure-mgmt-eventgrid
azure-mgmt-servicebus
azure-mgmt-logic
azure-mgmt-apimanagement
azure-mgmt-dns
azure-mgmt-msi
